Use new CSS environment variables: x/y/width/height (#12)
This changes the old CSS environment variables as follows:
| Old  | New |
| - | - |
| `titlebar-area-inset-left` | `titlebar-area-x` |
| `titlebar-area-inset-top`  | `titlebar-area-y` |
| `titlebar-area-inset-right` | `titlebar-area-width` |
| `titlebar-area-inset-bottom` | `titlebar-area-height` |

The motivation for this was that during development, we noticed that the use of `titlebar-area-inset-bottom` led to rough UI layouts of the pseudo "titlebar" region of the webpage when resizing the page vertically. One of the goals of this feature is a "native-like experience" for PWAs, the previous variables did not support that goal. 

The new variables take advantage of the fact that the height of the overlay does not change during resizes to provide a smoother UI during resize. Note: the height may change relative to the webpage when zooming, but since zooms usually occur stepwise, this should be less of an issue.