Update privacy with new fingerprinting risk (#20)
Previously we thought the fingerprinting risk was scoped to only expose the OS that the user is running (e.g. Mac title bars are shorter than Windows, and this would be exposed as the height of the overlay). However, we have since found that other OS-level and browser-level customizations are exposed by window controls overlay, but they are only exposed when a web app is installed and the feature is enabled.