Remove Open Questions section (#25)
This PR removes the Open Questions section and references to the section.

The open questions have already been discussed and rejected, so they are no longer needed. If we get enough feedback about any of the points posed in the Open Questions section, we can discuss adding them to the explainer/feature in the future.