Adjust xrefs, add dfn for event init dictionary
This fixes ReSpec xref config (`web-platform` is a profile and cannot appear in
the list of xref specs, unless it is the only value), adjusts cross-references
to `window`, and makes `[[TitleBarArea]]` an internal slot of the `Window`
interface.

To fix the Web IDL, this also introduces a definition for the dictionary
`WindowControlsOverlayGeometryChangeEventInit`. Note the exact definition of
that dictionary may need to be adjusted, this is just intended to make the
Web IDL valid so that the spec can be crawled and its IDL parsed by automated
tools.